requirements.txt
tensorflow>=2.3
numpy
matplotlib
pillow
scipy